+++ app/controllers/users/confirmations_controller.rb
@@ -17,7 +17,7 @@ module Users
     private
 
     def find_user
-      @user = User.not_confirmed.find_by(email: params[:email]&.strip&.downcase)
+      @user = User.not_confirmed.not_banned.find_by(email: params[:email]&.strip&.downcase)
       return if @user.present?
 
       failed_complete_confirmation