+++ app/controllers/users/recovery_controller.rb
@@ -25,7 +25,7 @@ module Users
     end
 
     def find_user
-      @user = User.find_by(email: params[:email]&.strip&.downcase)
+      @user = User.not_banned.find_by(email: params[:email]&.strip&.downcase)
       return if @user.present?
 
       failed_recovery